+++ ParticlePhaseSpace/DataLoaders.py
@@ -95,7 +95,7 @@ class _DataLoadersBase(ABC):
                                  f'\n{NaN_cols}')
 
         tot_mom = np.sqrt(self.data[self._columns['px']]**2 + self.data[self._columns['py']]**2 + self.data[self._columns['pz']]**2)
-        if not np.min(tot_mom)>0:
+        if not np.min(tot_mom) > 0:
             ind = tot_mom <= 0
             warnings.warn(f'{np.count_nonzero(ind)} particles have zero absolute momentum; this makes no sense and they will be removed')
 
