+++ ParticlePhaseSpace/__init__.py
@@ -1,7 +1,4 @@
-"""
-This library allows a user to plot a sine wave. It's fully sick.
-"""
-__version__ = '0.5.3'
+__version__ = '0.5.4'
 
 # from ParticlePhaseSpace._ParticlePhaseSpace import PhaseSpace
 from ParticlePhaseSpace.__unit_config__ import ParticlePhaseSpaceUnits, UnitSet
