### 2023-04-17 #5356 @matthewfeickert
> thanks! I'll have a think about that. I assume you want people I have never published with before?

Depends on how recently you've published with them, as per the [JOSS Conflict of Interest Policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). But if they don't have any conflict of interest as described there (or that they feel themselves) former co-authors are welcome to review your submission.
