### 2023-04-17 #5356 @bwheelz36
- saw a cool talk the other day from Andreas Moll who managed computing infrastructure at the australian synchrtron. I think he would have a good mix of python/ accelerator physics if he has time and we have no COI. https://github.com/portrain

struggling to think of too many people I wouldn't have at least a perceived COI right now sorry...
