+++ ParticlePhaseSpace/DataLoaders.py
@@ -97,7 +97,7 @@ class _DataLoadersBase(ABC):
         tot_mom = np.sqrt(self.data[self._columns['px']]**2 + self.data[self._columns['py']]**2 + self.data[self._columns['pz']]**2)
         if not np.min(tot_mom) > 0:
             ind = tot_mom <= 0
-            warnings.warn(f'{np.count_nonzero(ind)} particles have zero absolute momentum; this makes no sense and they will be removed')
+            warnings.warn(f'{np.count_nonzero(ind)} particles have zero absolute momentum.')
 
         # is every particle ID unique?
         if not len(self.data[self._columns['particle id']].unique()) == len(self.data[self._columns['particle id']]):
