### 2023-05-20 bwheelz36/ParticlePhaseSpace#146 @benjaminbolling
Paper comments in pull request
Comments on the paper itself, see https://github.com/bwheelz36/ParticlePhaseSpace/pull/145.
