### 2023-05-20 bwheelz36/ParticlePhaseSpace#146 @benjaminbolling
In Acknowledgements, a reference is made to "Image-X Institute". Please add more details on the institute (e.g. that it is part of the Faculty of Medicine and Health at the University of Sydney).
