### 2023-05-20 bwheelz36/ParticlePhaseSpace#147 @benjaminbolling
Fix description of README.md in examples/
The README in examples/ (https://bwheelz36.github.io/ParticlePhaseSpace/examples.html) could be extended a bit, or have a dot at the end. 🙂
