### 2023-05-20 bwheelz36/ParticlePhaseSpace#148 @benjaminbolling
pytest package missing
pytest package missing from [setup configuration](https://github.com/bwheelz36/ParticlePhaseSpace/blob/main/setup.cfg).
