### 2023-05-20 bwheelz36/ParticlePhaseSpace#149 @benjaminbolling
Automated test
Please add a python script that automatically invokes the test scripts with multiple datasets producing a more complete report that is readable directly from the terminal.
