### 2023-05-20 bwheelz36/ParticlePhaseSpace#151 @benjaminbolling
Paper: State of the field
The "State of the field" is mentioned under "Statement of need". The application should also be compared to other already existing applications that also calculate phase space, with the results also compared - not only the p2sat and postpic codes, which ParticlePhaseSpace is "built upon".

Regarding the statement "built upon", please elaborate a bit more.
