### 2023-05-20 #5375 @benjaminbolling
Initial review completed, very nice package and paper.

Issues reported and to be solved, from my point of view:

- https://github.com/bwheelz36/ParticlePhaseSpace/issues/146
- https://github.com/bwheelz36/ParticlePhaseSpace/issues/147
- https://github.com/bwheelz36/ParticlePhaseSpace/issues/148
- https://github.com/bwheelz36/ParticlePhaseSpace/issues/149
- https://github.com/bwheelz36/ParticlePhaseSpace/issues/150
- https://github.com/bwheelz36/ParticlePhaseSpace/issues/151
