+++ examples/README.md
@@ -1,1 +1,3 @@
-Note that these example notebooks are directly used to build the [documentation](https://bwheelz36.github.io/ParticlePhaseSpace/examples.html)
+This folder contains example notebooks, which are directly used to build the [documentation](https://bwheelz36.github.io/ParticlePhaseSpace/examples.html).
+
+If you would like to have an example notebook you wrote build into the docs, you have to place it in this folder, and then update the `docsrc/conf.py` and `examples.rst` in the docsrc folder.
