+++ docsrc/index.rst
@@ -7,6 +7,11 @@ aims to serve as a general purpose python library for importing, analysing, mani
 .. image:: figure_1.png
   :alt: Code structure
 
+The code can be installed from PyPi::
+
+    pip install particle_phase_space
+
+If you want to develop the code, there are some additional dependencies which are listed in the `dev_requirements.txt` file in the `github repository <https://github.com/bwheelz36/ParticlePhaseSpace>`_.
 
 .. toctree::
    :maxdepth: 2
