### 2023-05-22 bwheelz36/ParticlePhaseSpace#149 @bwheelz36
Hi @benjaminbolling

If you type `pytest` into the repository root (assuming it is installed as per #148 !) then it will collect and run all tests, generating the below report. Does this address this issue? If not could you point me to an example of what you mean?

![image](https://github.com/bwheelz36/ParticlePhaseSpace/assets/64300337/6e58cb83-3ad3-4ce4-9775-fcf54c13415e)
