### 2023-05-22 bwheelz36/ParticlePhaseSpace#151 @bwheelz36
Hi @benjaminbolling

> Regarding the statement "built upon", please elaborate a bit more.

So, what happened here was that I started working on this package independently, before discovering the [p2sat](https://github.com/lesnat/p2sat) package and realizing there was some substantial overlap. However, the p2sat package was also lacking some features like test cases, CI, and extension mechanisms. So I got in touch with Leo (author of that package and co-author on this paper) and essentially we figured out together which useful features in p2sat could be incorporated into this code.

So, this is the long story of what I mean when I say "built upon"! Let me know if you think I should use different wording or be more elaborate.

> The application should also be compared to other already existing applications that also calculate phase space, with the results also compared - not only the p2sat and postpic code

We have done heuristic "sanity check" comparisons with p2sat, but what specifically would you like to see here?
