### 2023-05-25 bwheelz36/ParticlePhaseSpace#148 @benjaminbolling
Concern resolved! Thanks, just note that in https://github.com/bwheelz36/ParticlePhaseSpace/blob/main/dev_requirements, the pytest-cov is included twice.
