### 2023-05-25 9d7b7ca
Update dev_requirements
remove double entry of pytest-cov as noted in #148
