### 2023-05-25 bwheelz36/ParticlePhaseSpace#149 @benjaminbolling
My bad, I did not install the dev_requirements meaning pytest could not initiate (lots of missing packages).

Installed dev_requirements in the conda env for this package, pytest working as intended. Many thanks!
