### 2023-05-25 bwheelz36/ParticlePhaseSpace#151 @benjaminbolling
Thanks for the explanation 🙂

I would though also prefer to see a comparison to e.g. TraceWin (https://www.dacm-logiciels.fr/tracewin) and Synergia (https://synergia.fnal.gov), if possible
