### 2023-05-25 bwheelz36/ParticlePhaseSpace#151 @bwheelz36
So both of these programs are for simulating accelerators. This code could in theory enable one to write new `DataLoaders` or `DataExporters` that would enable genetic particle phase space data to be be imported/exported between these codes - so it's not possible to 'compare'. Also TraceWin is commerical software that I can't access. Synergia it looks like I could access, but it would be a long process to build from scratch - and even then, I couldn't actually carry out a comparison, I could only write a DataLoader to load some data from Synergia. For this point - if you have a file I can test on, I would be happy to do it, but tbh I don't really want to have to build synergia from scratch and then learn how to use it!
