### 2023-05-30 bwheelz36/ParticlePhaseSpace#156 @ax3l
Additional Data Formats?
Thank you for the JOSS submission in https://github.com/openjournals/joss-reviews/issues/5375 .

I really like the support of the IAEA data loaders.

Based on the extended abstract and [linked motivating discussion in it](https://github.com/nrc-cnrc/EGSnrc/discussions/712), I was wondering:
I am personally curious if, for phase space data, the openPMD standard [[1]](https://github.com/openPMD) [[2]](https://www.openpmd.org) (disclaimer: I lead this effort) could be helpful as an additional input loader source? We have by now a relatively large selection of accelerator codes supporting [openPMD as their output](https://github.com/openPMD/openPMD-projects#scientific-simulations) and also try to use it more in experimental laser-plasma accelerator work.

If one were to implement another loader, how much work would be needed?

Note that the linked reference `Kuschel, S. (2022). Postpic.` https://github.com/skuschel/postpic implemented openPMD early on.
Minor correction: I think it should read `(2014)` as of the first release for this reference.

[1] https://github.com/openPMD
[2] https://www.openPMD.org
