### 2023-05-30 bwheelz36/ParticlePhaseSpace#157 @ax3l
Extended References on Toolkits & Standards
Thank you for the JOSS submission in https://github.com/openjournals/joss-reviews/issues/5375 .

This is question is related to #151 and #156.

## Recent Toolkits for Beam Analysis

Related to paper
> There are some additional open source codes providing similar functionality to this code; these include the p2sat (lesnat (2021)) and postpic (Kuschel (2022)) codes.

Over the recent years, a couple of our colleagues implemented new analysis packages that might be interesting to either be compatible with with cite as more recent related works. In particular, I am thinking of these packages:
- (SLAC) https://github.com/ChristopherMayes/openPMD-beamphysics ([docs here](https://christophermayes.github.io/openPMD-beamphysics), a lot of beam diagnostics)
- (LBNL, CFEL Hamburg University) https://github.com/openPMD/openPMD-viewer ([docs here](https://github.com/openPMD/openPMD-viewer/tree/dev/tutorials), incl. laser-plasma accelerator diagnostics)

## Accelerator Physics Codes

Additionally to the Wikipedia page linked as ` “Accelerator Physics Codes” (2022)` and maintained by our colleague David Sagan, maybe these catalogue that we created in the Consortium for Advanced Modeling of Particle Accelerators (CAMPA) are useful:

- https://github.com/campa-consortium/accelerator_simulation_codes / https://campa-consortium.github.io/accelerator_simulation_codes/

Related by biased, I do work on this set of accelerator modeling codes:
- https://blast.lbl.gov/codes/

The paper quotes
> Each of these programs will typically utilize their own unique format for saving and loading phase space data. This means that getting these programs to ‘speak’ to each other is generally a substantial amount of work.

I fully agree.
We hope to overcome this, since we actively standardize I/O with openPMD in many of these codes (see #156). Thus, the first sentence could maybe be rephrased slightly or cite projects like [openPMD (2015)](https://openpmd-api.readthedocs.io/en/latest/citation.html)?
