### 2023-05-30 bwheelz36/ParticlePhaseSpace#158 @ax3l
Data Sizes
Thank you for the JOSS submission in https://github.com/openjournals/joss-reviews/issues/5375 .

This is a follow-up question to #156.

In the design of this package, what are the envisioned data sizes for phase space data to be processed? Up to the size of a laptop RAM/single node?

I was looking at
  https://bwheelz36.github.io/ParticlePhaseSpace/new_data_loader.html

and am wondering if not most of the operations here are map-reduce operations and could be implemented to stream over arbitrary data sizes, e.g., if large simulation data is being processed?

I did some experiments on processing such data with Dask:
https://github.com/openPMD/openPMD-api/pull/963#issuecomment-873350174

and wonder if something similar could be used as the backend here to scale up? :rocket:
