### 2023-05-30 #5375 @ax3l
I hope that is not a silly question :sweat_smile:

For the question
> Contribution and authorship: Has the submitting author (@bwheelz36) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

I cannot immediately spot the contributions of Leo Esnault @lesnat in https://github.com/bwheelz36/ParticlePhaseSpace/graphs/contributors
Can you please comment on the authors contributions? (I must have missed it, I see that he developed `p2sat` in related works.)
