### 2023-05-30 aba2ee6
Readme: Related Packages
Related to #157, but might also be cited in JOSS :)
