+++ README.md
@@ -35,4 +35,6 @@ or request support
 Below is a (almost certainly incomplete) list of other packages I have come accross for analysis of particle phase space data:
 
 - [p2sat](https://github.com/lesnat/p2sat)
-- [postpic](https://github.com/skuschel/postpic) 
+- [postpic](https://github.com/skuschel/postpic)
+- [openPMD-beamphysics](https://christophermayes.github.io/openPMD-beamphysics)
+- [openPMD-viewer](https://github.com/openPMD/openPMD-viewer)
