### 2023-05-31 c4126bd
Doc: PyPI link to RTD
There is a cool feature to cross-link the user-facing documentation
(and other URLs) in PyPI, too. This saves users a hop over the source
repo and might be useful for completion.
