+++ setup.cfg
@@ -10,6 +10,7 @@ long_description_content_type = text/markdown
 url = https://github.com/bwheelz36/ParticlePhaseSpace
 project_urls =
     Bug Tracker = https://github.com/bwheelz36/ParticlePhaseSpace/issues
+    Documentation = https://bwheelz36.github.io/ParticlePhaseSpace
 classifiers =
     Programming Language :: Python :: 3
     Operating System :: OS Independent
