### 2023-05-31 #5375 @bwheelz36
Hi @ax3l, no problem :-)
Leo's contribution is discussed in https://github.com/bwheelz36/ParticlePhaseSpace/issues/151;

"...I got in touch with Leo (author of that package and co-author on this paper) and essentially we figured out together which useful features in p2sat could be incorporated into this code."
Essentially some of the features are based on his earlier work on p2sat; he also contributed by reviewing and making suggestions around the code architecture and capability.
