### 2023-05-31 bwheelz36/ParticlePhaseSpace#156 @ax3l
Hi @bwheelz36 , sorry for the edit in my original message.

I added a few example files and a probably four liner to load data via an edit :)
