### 2023-06-01 bwheelz36/ParticlePhaseSpace#156 @bwheelz36
Hi @ax3l

That's all good - given there is a defined open dataset format, it absolutely makes sense that this package should support it.

Having said that - I'm a bit confused tbh. I'm trying to run the [first read](https://openpmd-api.readthedocs.io/en/latest/usage/firstread.html) example from the openpmd-api site with the following code:

```python
import openpmd_api as io
series = io.Series( "data%T.h5", io.Access.read_only)
```

I pointed this code to each of the three examples `example-2d', `example-3d', `example-thetaMode` - (it is actually not that clear from the example that this is what you are supposed to do?). In each case the data loads, but there is no information in the 'iterations' attribute?
