### 2023-06-02 26ae4b7
update paper with new refs as per #157
