+++ paper.md
@@ -28,7 +28,7 @@ In accelerator particle physics, a description of the positions and directions o
 
 # Statement of need
 
-Although the use of phase space data is well established, there is no consistent implementation of phase space data between different programs, as discussed at length in @tessier_proposal_2021. To appreciate why this is an issue, one must understand that in a typical accelerator workflow, it is common to utilize several different programs to simulate different phases of particle transport. For any given simulation task, there are many different simulation programs one can use; an incomplete list is supplied in the wikipedia article @noauthor_accelerator_2022.  Each of these programs will typically utilize their own unique format for saving and loading phase space data. This means that getting these programs to 'speak' to each other generally requires a substantial amount of work. In addition, the fragmented data format has led to fragmented analysis frameworks, with different fields and research groups tending to develop their own, generally in-house, code for handling phase space data.
+Although the use of phase space data is well established, there is no consistent implementation of phase space data between different programs, as discussed at length in @tessier_proposal_2021. To appreciate why this is an issue, one must understand that in a typical accelerator workflow, it is common to utilize several different programs to simulate different phases of particle transport. For any given simulation task, there are many different simulation programs one can use; an incomplete list is supplied in the wikipedia article @noauthor_accelerator_2022 and the campa consortium data base @noauthor_acceleration_nodate.  Each of these programs will typically utilize their own unique format for saving and loading phase space data. This means that getting these programs to 'speak' to each other generally requires a substantial amount of work. The openPMD initiative  aims to solve this problem by defining a common open-source data standard @openPMDstandard. However, many codes have yet to adopt this standard, and some commercial codes in particular may never do so. The fragmented nature of phase space data formats has in turn led to fragmented analysis frameworks, with different fields and research groups tending to develop their own, often in-house, code for handling phase space data.
 
 `ParticlePhaseSpace` aims to solve these issues by providing well documented, extensible mechanisms for the import and export of data in different formats, as well as a library of methods for visualizing, manipulating, characterizing and analyzing phase space data. There are many examples in the scientific literature where ParticlePhaseSpace would be useful, for example; @st_aubin_integrated_2010, @whelan_novel_2016, @lesnat_particle_2021. The basic code hierarchy is shown in \autoref{figure 1}. 
 
@@ -38,7 +38,7 @@ Import/Export of different data formats is facilitated through the use of Abstra
 
 ![Examples of plots from ParticlePhaseSpace. A) Multi-particle energy histogram. B) 2D intensity histogram of gamma particles, C) Trace-space in X of gamma particles\label{figure 2}](figure_2.png)
 
-There are some additional open source codes providing similar functionality to this code; these include the p2sat (@lesnat_particle_2021) and postpic (@kuschel_postpic_2022) codes. `ParticlePhaseSpace` builds on these codes with extension mechanisms, a testing framework with continuous integration, multi-particle support in the same Phase Space object, and automatic code documentation.  In summary, `ParticlePhaseSpace` provides an open source, well documented, well tested, and extensible library enabling import/analysis/export of PhaseSpace data of arbitrary format.
+There are some additional open source codes providing similar functionality to this code; these include the p2sat (@lesnat_particle_2021) and postpic (@kuschel_postpic_2022) codes. The codes @noauthor_openpmd-viewer_2023 and @mayes_openpmd-beamphysics_2023 have been specifically designed to work with the openPMD standard. `ParticlePhaseSpace` provides an additional tool in this landscape and provides extension mechanisms for loading and exporting arbitrary data formats, a testing framework with continuous integration, multi-particle support in the same Phase Space object, many useful plotting and analysis routines, and automatic code documentation.
 
 # Acknowledgements
 
