+++ paper.bib
@@ -143,3 +143,58 @@
 	note = {Publisher: Wiley Online Library},
 	pages = {7623--7637},
 }
+
+@misc{mayes_openpmd-beamphysics_2023,
+	title = {{openPMD}-beamphysics},
+	copyright = {Apache-2.0},
+	url = {https://github.com/ChristopherMayes/openPMD-beamphysics},
+	abstract = {Tools for analyzing and viewing particle data in the openPMD standard, extension beamphysics.},
+	urldate = {2023-06-02},
+	author = {Mayes, Christopher},
+	month = jan,
+	year = {2023},
+	note = {original-date: 2019-06-17T17:45:15Z},
+}
+
+@misc{noauthor_openpmd-viewer_2023,
+	title = {{openPMD}-viewer},
+	url = {https://github.com/openPMD/openPMD-viewer},
+	abstract = {:snake: Python visualization tools for openPMD files},
+	urldate = {2023-06-02},
+	publisher = {openPMD},
+	month = feb,
+	year = {2023},
+	note = {original-date: 2015-10-08T16:22:47Z},
+	keywords = {community, jupyter-notebook, openpmd, openscience, research, visualization},
+}
+
+@misc{openPMDstandard,
+  author       = {Huebl, Axel and
+                  Lehe, R{\'e}mi and
+                  Vay, Jean-Luc and
+                  Grote, David P. and
+                  Sbalzarini, Ivo and
+                  Kuschel, Stephan and
+                  Sagan, David and
+                  Mayes, Christopher and
+                  P{\'e}rez, Fr{\'e}d{\'e}ric and
+                  Koller, Fabian and
+                  Bussmann, Michael},
+  title        = {{openPMD: A meta data standard for particle and mesh based data}},
+  year         = 2015,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.591699},
+  url          = {https://www.openPMD.org},
+  howpublished = {https://github.com/openPMD}
+}
+
+@misc{noauthor_acceleration_nodate,
+	title = {Acceleration {Simulation} {Codes}},
+	url = {https://campa-consortium.github.io/accelerator_simulation_codes/},
+	abstract = {This is a community database of accelerator simulation codes.},
+	language = {de},
+	urldate = {2023-06-02},
+	journal = {Acceleration Simulation Codes},
+	note = {Section: Research},
+	file = {Snapshot:/home/brendan/Zotero/storage/45NZTV3V/accelerator_simulation_codes.html:text/html},
+}
