### 2023-06-05 bwheelz36/ParticlePhaseSpace#156 @ax3l
Hi @bwheelz36,

Thanks for trying the example datasets!
The iterations concept is explained here:
https://openpmd-api.readthedocs.io/en/latest/usage/concepts.html

> there is no information in the 'iterations' attribute?

Please let me know if you have more questions on this in case I missed the point of the question :)

Once you open a data `Series`, you can loop over available iterations in it, read the data
- fields: https://openpmd-api.readthedocs.io/en/latest/usage/firstread.html or (for our case here)
- particles: https://github.com/openPMD/openPMD-api/blob/0.15.1/examples/2_read_serial.cpp#L48-L67
  - or as data frame (preferred here): https://github.com/openPMD/openPMD-api/blob/0.15.1/examples/11_particle_dataframe.py#L30-L34
