### 2023-06-06 bwheelz36/ParticlePhaseSpace#156 @bwheelz36
Hi @ax3l

Ok,  here's an end to end example of what I tried. Maybe I'm doing something extremely stupid...

in a terminal:

```bash
# inside a fresh virtual environment
git clone https://github.com/openPMD/openPMD-example-datasets.git
cd openPMD-example-datasets
tar -zxvf example-2d.tar.gz
tar -zxvf example-3d.tar.gz
tar -zxvf example-thetaMode.tar.gz

pip install openpmd-api
python  # enter python session
```

inside python:
```python
import openpmd_api as io

data_loc = "example-2d/hdf5/data%T.h5"
s = io.Series(data_loc, io.Access.read_only)
```
Here's the explorer view of s; it appears to simply have nothing in it?

![image](https://github.com/bwheelz36/ParticlePhaseSpace/assets/64300337/50de5555-05f1-4945-996c-fe7571e40d6c)
