+++ paper.bib
@@ -112,7 +112,7 @@
 	url = {https://github.com/lesnat/p2sat},
 	abstract = {Particle Phase Space Analysis Toolkit},
 	urldate = {2023-01-23},
-	author = {lesnat},
+	author = {Esnault, Leo},
 	month = feb,
 	year = {2021},
 	note = {original-date: 2018-04-16T19:24:48Z},
@@ -146,6 +146,7 @@
 
 @misc{mayes_openpmd-beamphysics_2023,
 	title = {{openPMD}-beamphysics},
+	author = {}
 	copyright = {Apache-2.0},
 	url = {https://github.com/ChristopherMayes/openPMD-beamphysics},
 	abstract = {Tools for analyzing and viewing particle data in the openPMD standard, extension beamphysics.},
@@ -158,6 +159,7 @@
 
 @misc{noauthor_openpmd-viewer_2023,
 	title = {{openPMD}-viewer},
+	author = {Lehe, Remi and Huebl, Axel and Jalas, Sören and Thévenet, Maxence et. al.}
 	url = {https://github.com/openPMD/openPMD-viewer},
 	abstract = {:snake: Python visualization tools for openPMD files},
 	urldate = {2023-06-02},
@@ -169,17 +171,7 @@
 }
 
 @misc{openPMDstandard,
-  author       = {Huebl, Axel and
-                  Lehe, R{\'e}mi and
-                  Vay, Jean-Luc and
-                  Grote, David P. and
-                  Sbalzarini, Ivo and
-                  Kuschel, Stephan and
-                  Sagan, David and
-                  Mayes, Christopher and
-                  P{\'e}rez, Fr{\'e}d{\'e}ric and
-                  Koller, Fabian and
-                  Bussmann, Michael},
+  author       = {Consortium for Advanced Modeling of Particle Accelerators (CAMPA)},
   title        = {{openPMD: A meta data standard for particle and mesh based data}},
   year         = 2015,
   publisher    = {Zenodo},
