+++ paper.bib
@@ -146,7 +146,6 @@
 
 @misc{mayes_openpmd-beamphysics_2023,
 	title = {{openPMD}-beamphysics},
-	author = {}
 	copyright = {Apache-2.0},
 	url = {https://github.com/ChristopherMayes/openPMD-beamphysics},
 	abstract = {Tools for analyzing and viewing particle data in the openPMD standard, extension beamphysics.},
