+++ paper.bib
@@ -158,7 +158,7 @@
 
 @misc{noauthor_openpmd-viewer_2023,
 	title = {{openPMD}-viewer},
-	author = {Lehe, Remi and Huebl, Axel and Jalas, Sören and Thévenet, Maxence et. al.}
+	author = {Lehe, Remi and Huebl, Axel and Jalas, Sören and Thévenet, Maxence et. al.},
 	url = {https://github.com/openPMD/openPMD-viewer},
 	abstract = {:snake: Python visualization tools for openPMD files},
 	urldate = {2023-06-02},
@@ -172,7 +172,7 @@
 @misc{openPMDstandard,
   author       = {Consortium for Advanced Modeling of Particle Accelerators (CAMPA)},
   title        = {{openPMD: A meta data standard for particle and mesh based data}},
-  year         = 2015,
+  year         = {2015},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.591699},
   url          = {https://www.openPMD.org},
