### 2023-06-06 bwheelz36/ParticlePhaseSpace#157 @bwheelz36
@ax3l, I updated as per your comments in [this commit](https://github.com/bwheelz36/ParticlePhaseSpace/commit/87f56edb54ce8f2ad8d0f66532466a76cfe2c53d) (sorry I forgot to tag this issue).
I had to make a few more minor changes to fix some typos, but that commit contains the bulk of the corrections
