### 2023-06-06 786a604
add limitations to docs as per #158
