### 2023-06-06 bwheelz36/ParticlePhaseSpace#158 @bwheelz36
Hey @ax3l - I won't realistically be addressing this concern any time soon, but I think it is a very valid point - as such, I've added a [page](https://bwheelz36.github.io/ParticlePhaseSpace/limitations.html) to the docs called limitations, which details this as what I think is the major limitation of this code at present...
