### 2023-06-10 bwheelz36/ParticlePhaseSpace#151 @benjaminbolling
Sorry for the long waiting time for my reply.

Thanks for clarifying - it is probably my background in accelerator physics that made me want to attempt to compare the data to some software I am used to, but indeed, the applications are more for simulating accelerators rather than working with generic particle phase space data.
