### 2023-06-10 bwheelz36/ParticlePhaseSpace#163 @benjaminbolling
Paper: Add limitations/future improvements
As described in https://github.com/bwheelz36/ParticlePhaseSpace/issues/158 regarding data sizes, a page has been added describing limitations being handling data that can fit inside e.g. a laptop's internal memory.

As also explained in the page, libraries such as DASK may enable using this library of distributed resources.

Could this be perhaps mentioned in the paper as well, as current limitations of the software and being a potential future improvement - together with reference to DASK?
