### 2023-06-10 #5375 @benjaminbolling
All previous issues reported by me are closed, new opened following @ax3l's issue [158](https://github.com/bwheelz36/ParticlePhaseSpace/issues/158):

- https://github.com/bwheelz36/ParticlePhaseSpace/issues/163
