+++ paper.md
@@ -38,7 +38,7 @@ Import/Export of different data formats is facilitated through the use of Abstra
 
 ![Examples of plots from ParticlePhaseSpace. A) Multi-particle energy histogram. B) 2D intensity histogram of gamma particles, C) Trace-space in X of gamma particles\label{figure 2}](figure_2.png)
 
-There are some additional open source codes providing similar functionality to this code; these include the p2sat (@lesnat_particle_2021) and postpic (@kuschel_postpic_2022) codes. The codes @noauthor_openpmd-viewer_2023 and @mayes_openpmd-beamphysics_2023 have been specifically designed to work with the openPMD standard. `ParticlePhaseSpace` provides an additional tool in this landscape and provides extension mechanisms for loading and exporting arbitrary data formats, a testing framework with continuous integration, multi-particle support in the same Phase Space object, many useful plotting and analysis routines, and automatic code documentation.
+There are some additional open source codes providing similar functionality to this code; these include the p2sat (@lesnat_particle_2021) and postpic (@kuschel_postpic_2022) codes. The codes @noauthor_openpmd-viewer_2023 and @mayes_openpmd-beamphysics_2023 have been specifically designed to work with the openPMD standard. `ParticlePhaseSpace` provides an additional tool in this landscape and provides extension mechanisms for loading and exporting arbitrary data formats, a testing framework with continuous integration, multi-particle support in the same Phase Space object, many useful plotting and analysis routines, and automatic code documentation. The major limitation of the code at the time of writing is the ability to handle data larger than memory. Future versions could address this limitation by adding well supported 'chunking' mechanisms or using libraries such as dask @DASK to enable distributed operations. 
 
 # Acknowledgements
 
