+++ paper.bib
@@ -181,6 +181,7 @@
 
 @misc{noauthor_acceleration_nodate,
 	title = {Acceleration {Simulation} {Codes}},
+	year = {2022},
 	url = {https://campa-consortium.github.io/accelerator_simulation_codes/},
 	abstract = {This is a community database of accelerator simulation codes.},
 	language = {de},
@@ -188,4 +189,11 @@
 	journal = {Acceleration Simulation Codes},
 	note = {Section: Research},
 	file = {Snapshot:/home/brendan/Zotero/storage/45NZTV3V/accelerator_simulation_codes.html:text/html},
+}
+
+@Manual{DASK,
+  title = {Dask: Library for dynamic task scheduling},
+  author = {{Dask Development Team}},
+  year = {2016},
+  url = {https://dask.org},
 }
