### 2023-06-25 #5375 @ax3l
Thank you so much for the update to the paper @bwheelz36.

I only see one issue in the openPMD (2015) reference on line 28f.: CAMPA in the US and other projects in Europe support it, but the authorship is actually me :sweat_smile:

```
@misc{openPMDstandard,
  author       = {Huebl, Axel and
                  Lehe, R{\'e}mi and
                  Vay, Jean-Luc and
                  Grote, David P. and
                  Sbalzarini, Ivo and
                  Kuschel, Stephan and
                  Sagan, David and
                  P{\'e}rez, Fr{\'e}d{\'e}ric and
                  Koller, Fabian and
                  Bussmann, Michael},
  title        = {{openPMD: A meta data standard for particle and mesh based data}},
  year         = 2015,
  publisher    = {Zenodo},
  doi          = {10.5281/zenodo.591699},
  url          = {https://www.openPMD.org},
  howpublished = {https://github.com/openPMD}
}
```

Classic reviewer number 3 comment, sorry for that :)
