### 2023-06-25 bwheelz36/ParticlePhaseSpace#157 @ax3l
Thanks a lot @bwheelz36. Please excuse the delay.

I noticed one detail and commented a suggestion in
- https://github.com/openjournals/joss-reviews/issues/5375#issuecomment-1606260053
