### 2023-06-25 #5375 @ax3l
Please excuse my delay in finalizing my review.
I finalized the functional checks and all questions I raised have been answered and/or document and PRs been merged. I also also completed the functional tests of the software and it is satisfactory to me with the documented RTD pages.

Again, great work @bwheelz36 and looking forward to continue interacting together on future integrations :star_struck:

The linked issue above with the citation authorship is the last part I would like to see changed in the manuscript and I have no futher repo change requests.
