### 2023-06-25 ec7f655
openPMD-standard: Fix Authors
X-ref:
- https://github.com/bwheelz36/ParticlePhaseSpace/commit/87f56edb54ce8f2ad8d0f66532466a76cfe2c53d#r119609835
- https://github.com/openjournals/joss-reviews/issues/5375#issuecomment-1606260053
- https://github.com/bwheelz36/ParticlePhaseSpace/issues/157#issuecomment-1606261321

Close #157
