+++ paper.bib
@@ -170,9 +170,19 @@
 }
 
 @misc{openPMDstandard,
-  author       = {Consortium for Advanced Modeling of Particle Accelerators (CAMPA)},
+  author       = {Huebl, Axel and
+                  Lehe, R{\'e}mi and
+                  Vay, Jean-Luc and
+                  Grote, David P. and
+                  Sbalzarini, Ivo and
+                  Kuschel, Stephan and
+                  Sagan, David and
+                  Mayes, Christopher and
+                  P{\'e}rez, Fr{\'e}d{\'e}ric and
+                  Koller, Fabian and
+                  Bussmann, Michael},
   title        = {{openPMD: A meta data standard for particle and mesh based data}},
-  year         = {2015},
+  year         = 2015,
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.591699},
   url          = {https://www.openPMD.org},
