+++ paper.bib
@@ -182,7 +182,7 @@
                   Koller, Fabian and
                   Bussmann, Michael},
   title        = {{openPMD: A meta data standard for particle and mesh based data}},
-  year         = 2015,
+  year         = {2015},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.591699},
   url          = {https://www.openPMD.org},
