+++ paper.bib
@@ -191,7 +191,7 @@
 
 @misc{noauthor_acceleration_nodate,
 	title = {Acceleration {Simulation} {Codes}},
-        author = {{Consortium for Advanced Modeling of Particle Accelerators (CAMPA)}},
+	author = {{Consortium for Advanced Modeling of Particle Accelerators (CAMPA)}},
 	year = {2022},
 	url = {https://campa-consortium.github.io/accelerator_simulation_codes/},
 	abstract = {This is a community database of accelerator simulation codes.},
