+++ paper.bib
@@ -190,7 +190,7 @@
 }
 
 @misc{noauthor_acceleration_nodate,
-	title = {Acceleration {Simulation} {Codes}},
+	title = {{Accelerator Simulation Codes}},
 	author = {{Consortium for Advanced Modeling of Particle Accelerators (CAMPA)}},
 	year = {2022},
 	url = {https://campa-consortium.github.io/accelerator_simulation_codes/},
