+++ paper.bib
@@ -197,7 +197,7 @@
 	abstract = {This is a community database of accelerator simulation codes.},
 	language = {de},
 	urldate = {2023-06-02},
-	journal = {Acceleration Simulation Codes},
+	journal = {Accelerator Simulation Codes},
 	note = {Section: Research},
 	file = {Snapshot:/home/brendan/Zotero/storage/45NZTV3V/accelerator_simulation_codes.html:text/html},
 }
