+++ paper.bib
@@ -127,7 +127,7 @@
 	urldate = {2023-03-01},
 	author = {Kuschel, Stephan},
 	month = oct,
-	year = {2022},
+	year = {2014},
 	note = {original-date: 2014-06-22T20:55:11Z},
 	keywords = {python, particle-in-cell, physics, hacktoberfest, particles, physics-analysis, pic, plasma, post-processor, science, simulation, simulations},
 }
