### 2023-06-25 9c3108e
openPMD-viewer (2016)
The first functional releases were published in 2016.
https://github.com/openPMD/openPMD-viewer/releases/tag/0.2.0
