+++ paper.bib
@@ -164,7 +164,7 @@
 	urldate = {2023-06-02},
 	publisher = {openPMD},
 	month = feb,
-	year = {2023},
+	year = {2016},
 	note = {original-date: 2015-10-08T16:22:47Z},
 	keywords = {community, jupyter-notebook, openpmd, openscience, research, visualization},
 }
