### 2023-07-05 #5375 @matthewfeickert
```
...
MISSING DOIs

- 10.1002/mp.15887 may be a valid DOI for title: Bayesian optimization to design a novel x-ray shaping device
...
```

@bwheelz36 Can you add https://doi.org/10.1002/mp.15887 as the DOI for "Bayesian optimization to design a novel x-ray shaping device"?
