+++ paper.bib
@@ -138,7 +138,7 @@
 	volume = {49},
 	number = {12},
 	journal = {Medical Physics},
-	doi = {https://doi.org/10.1002/mp.15887},
+	doi = {10.1002/mp.15887},
 	author = {Whelan, Brendan and Trovati, Stefania and Wang, Jinghui and Fahrig, Rebecca and Maxim, Peter G. and Hanuka, Adi and Shumail, Muhammad and Tantawi, Sami and Merrick, Julian and Perl, Joseph},
 	year = {2022},
 	note = {Publisher: Wiley Online Library},
