### 2023-07-07 bwheelz36/ParticlePhaseSpace#156 @franzpoeschel
For this, see my comment [here](https://github.com/openPMD/openPMD-api/issues/1375#issuecomment-1438759626):

>  The string representations of many classes are counterintuitive and have led to confusion, e.g. `series.iterations` printed will look as if it is empty

I guess that this issue is proved again..
The data is there, it just does not look like it:

```
>>> import openpmd_api as io
>>> s = io.Series("data%T.h5", io.Access.read_only)
>>> s.iterations
<openPMD.Attributable with '0' attributes>
>>> [index for index in s.iterations]
[255, 260, 265, 270, 275, 280, 285, 290, 295, 300, 305, 310, 315, 320, 325, 330, 335, 340, 345, 350, 355, 360, 365, 370, 375, 380, 385, 390, 395, 400]
```
