### 2023-07-14 bwheelz36/ParticlePhaseSpace#156 @franzpoeschel
Fixed in https://github.com/openPMD/openPMD-api/pull/1476
