### 2023-07-17 bwheelz36/ParticlePhaseSpace#156 @ax3l
Thank you for updating the representation strings, @franzpoeschel! This will be shipped with the next patch release, `0.15.2`.

@bwheelz36 for your example above, all looks good and you can keep exploring what is inside the data series `s` [like this](https://github.com/openPMD/openPMD-api/tree/0.15.1#python):
```py
for k_i, i in s.iterations.items():
    print("Iteration: {0}".format(k_i))

    for k_p, p in i.particles.items():
        print("  Particle species '{0}':".format(k_p))
```

inside the particle species `p` is then a record component that is a key-value pair of a string + record component, which can be [accessed like a numpy array](https://openpmd-api.readthedocs.io/en/0.15.1/usage/firstread.html#register-chunk), e.g., `u_x = p["momentum"]["x"][()]` - note that `s.flush()` will fill the array `u_x` with actual data.

Even easier is the [access as a data frame](https://openpmd-api.readthedocs.io/en/0.15.1/usage/examples.html#python), as in  the [11_particle_dataframe.py](https://github.com/openPMD/openPMD-api/blob/dev/examples/11_particle_dataframe.py) example:
```py
for k_i, i in s.iterations.items():
    print("Iteration: {0}".format(k_i))

    for k_p, p in i.particles.items():
        print("  Particle species '{0}':".format(k_p))
        df = p.to_df()
        print(df)
```
