### 2023-08-16 bwheelz36/ParticlePhaseSpace#156 @bwheelz36
Hi @ax3l - the first loop you posted above helps yes - it is clear there is some data there! in that example, doing `p.to_df()` gives a dataframe which would facilitate close to one-to-one read in to ParticlePhaseSpace.

the second loop crashes with `AttributeError: 'int' object has no attribute 'particles'`. I added a line `if hasattr(i, 'particles'):` however this was never entered...

Can I make sure I understand the intent behind iterations - each iteration would represent for instance a time interval?
