### 2023-08-16 bwheelz36/ParticlePhaseSpace#156 @franzpoeschel
> the second loop crashes with `AttributeError: 'int' object has no attribute 'particles'`. I added a line `if hasattr(i, 'particles'):` however this was never entered...

I think that there is a slight bug in the second loop, try this one:

```
for it_index, it in s.iterations.items():
    for p in it.particles:
        df = p.to_df()
        print(df)
```
