### 2023-08-30 #5375 @matthewfeickert
https://github.com/bwheelz36/ParticlePhaseSpace/pull/165 addresses minor typos, formatting, and missing DOIs. If there are questions on this please do not hesitate to ask &mdash; these are not mandates, they are suggestions.

---

@bwheelz36 Thanks for your patience. There isn't an excuse for the delays you've waited through, as I should have transitioned this to an EiC to finish given that I was effectively out of office and internet contact given unexpected housing saftey issues.
