### 2023-08-30 #5375 @matthewfeickert
@bwheelz36 all that's left to do now before publication is to ensure that there is a long term public archive of the code that was reviewed, in this case [`ParticlePhaseSpace` `v0.5.4`](https://github.com/bwheelz36/ParticlePhaseSpace/releases/tag/v0.5.4).

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [an (optional) GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

Once you have a DOI for the archive please just report them here and I'll have @editorialbot update the version listed in the review and add the archive.

Let me know if you have any questions. :+1:
