### 2023-09-01 #5375 @matthewfeickert
> Can you also revise the Zenodo archive metadata through the website to have the authors be the authors listed on the paper (this is a minor thing that we ask for all JOSS papers)?

@bwheelz36 It is already Friday evening in Australia and so I don't expect that you will have time to get to this before the next week (totally reasonable). I'll go ahead and still recommend publication to the Editor in Chief so that this can get put on their radar to look at once the author metadata for the archive gets updated. :+1:
