### 2023-09-02 #5375 @bwheelz36
> > Great, done: https://zenodo.org/record/8299964
>
> @bwheelz36 Excellent. Can you also revise the Zenodo archive metadata through the website to have the authors be the authors listed on the paper (this is a minor thing that we ask for all JOSS papers)? After that I'll recommend the EiC trigger publication.

done :-)
